owasp_gcp_scanner/requirements.txt
google-api-python-client==2.134.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-cloud-storage==2.18.2
PyYAML==6.0.2
tqdm==4.66.5